refactor: centralize tools_condition_worker and add serpapi for google search
