+++ pyproject.toml
@@ -38,6 +38,7 @@ asgiref = "^3.9.1"
 ragas = "^0.3.1"
 pillow = "^11.3.0"
 rapidfuzz = "^3.13.0"
+google-search-results = "^2.4.2"
 
 [build-system]
 requires = ["poetry-core>=2.0.0,<3.0.0"]
