+++ src/agents/search/tool_google.py
@@ -1,28 +1,26 @@
 import re
+from langchain_core.tools.base import InjectedToolCallId
 from typing import Annotated
+from langgraph.prebuilt import InjectedState
 from langchain_core.tools import StructuredTool
 from serpapi import GoogleSearch
-from langgraph.prebuilt import InjectedState
-from langchain_core.tools.base import InjectedToolCallId
 from pydantic import BaseModel, Field
 import os
 from dotenv import load_dotenv
+from agents.utils import limit_calls
 
 # Load SerpAPI key from .env
 load_dotenv()
 SERPAPI_API_KEY = os.getenv("SERP_API_KEY")
 
 
-# ------------------------
-# Input Schema for Search
-# ------------------------
 class GoogleSearchSchema(BaseModel):
+    tool_call_id: Annotated[str, InjectedToolCallId]
+    state: Annotated[dict, InjectedState]
     query: str = Field(..., description="The search query to look up on Google.")
 
 
-# ------------------------
-# Search Tool Logic
-# ------------------------
+@limit_calls(max_calls=10)
 def perform_google_search(
     tool_call_id: Annotated[str, InjectedToolCallId],
     state: Annotated[dict, InjectedState],
@@ -56,9 +54,6 @@ def perform_google_search(
     return {"answer": f"Top 3 search results for: {query}", "results": output}
 
 
-# ------------------------
-# Define Tool
-# ------------------------
 google_search_tool = StructuredTool(
     name="google_search_tool",
     description="Uses Google to search for information and returns top 3 results with clean text.",
@@ -66,5 +61,4 @@ google_search_tool = StructuredTool(
     func=perform_google_search,
 )
 
-# Expose as toolset for worker
 search_toolset = [google_search_tool]
