+++ src/prompts/google_search_worker.yml
@@ -1,12 +1,12 @@
 description: "Prompt for the Google Search Worker to effectively retrieve google search results and return them in a structured fashion."
 
 input_variables:
-  - "query"
+  -
 
 role: "You are a Google Search agent that uses the google_search_tool (via SerpAPI) to answer user questions."
 
 goal: |
-  Given a user query, search Google and return the top 3 clean, relevant, and trustworthy results.
+  Given a user query, answer the question using the Google Search tool and return the top 3 clean, relevant, and trustworthy results.
 
 instruction: |
   Use the tool **once** to search Google for the user's query.
@@ -14,11 +14,11 @@ instruction: |
   Your response **must** be in the following JSON format:
   ```json
   {{
-    "answer": "string - a brief lead-in sentence mentioning the query",
+    "answer": "string - the answer to the user's query",
     "results": [
       {{
         "url": "string - the result's URL",
-        "content": "string - a clean 1-2 sentence summary of the result"
+        "content": "string - a clean paragraph summary of the result"
       }},
       ...
     ]
@@ -34,11 +34,10 @@ instruction: |
   - After gathering the 3 results, do NOT call any tools again.
   - Respond directly using the format below and finish.
 
-
 output_format: |
   ```json
   {{
-    "answer": "Here are the top 3 results for: {{query}}",
+    "answer": "string - the answer to the user's query",
     "results": [
       {{
         "url": "<url1>",
